Rename folder: state -> providers
